FX-26, Workflow fixes, binary upload fixes
git-svn-id: http://svn.flexive.org/repository/flexive/trunk@97 d35d4baa-7912-0410-b0ae-edacad209e15

+++ src/framework/java/com/flexive/core/security/UserTicketImpl.java
@@ -142,7 +142,7 @@ public class UserTicketImpl implements UserTicket, Serializable {
     /**
      * {@inheritDoc}
      */
-    public boolean isInGroup(int group) {
+    public boolean isInGroup(long group) {
         return FxArrayUtils.containsElement(groups, group);
     }
